requirements.txt
numpy
python-can
pyserial
matplotlib
pandas
tqdm
pyqtgraph
pyqt5
NeuroLocoMiddleware